projeto_back_end/pyproject.toml
[project]
name = "pt-manager"
version = "0.1.0"
description = "Single-user Personal Trainer client management API (FastAPI + SQLModel + SQLite)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlmodel>=0.0.16",
    "pydantic-settings>=2.2",
    "psycopg2-binary>=2.9.9",  # Driver PostgreSQL
    "apscheduler>=3.10.4",      # Agendamento de tarefas
    "requests>=2.31.0",          # HTTP client para WhatsApp API
    "python-multipart>=0.0.9",   # Para uploads de arquivos (se necessário)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "httpx>=0.25",  # Para testes de API
    "ruff>=0.1",    # Linter e formatter
]

[tool.uvicorn]
factory = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []